exemplo-java-gradle/api/src/main/java/com/example/demo/service/UserDetailsService.java
package com.example.demo.service;

import com.example.demo.data.UserDetailsData;
import com.example.demo.domain.User;
import com.example.demo.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Component;

import java.util.Optional;

@Component
public class UserDetailsService implements org.springframework.security.core.userdetails.UserDetailsService {

    @Autowired
    private UserRepository repository;

    @Override
    public UserDetails loadUserByUsername(String login) throws UsernameNotFoundException {
        Optional<User> user = repository.findByLogin(login);
        if (user.isEmpty()){
            throw new UsernameNotFoundException(String.format("Usuário %s não encontrado", login));
        }

        return new UserDetailsData(user);
    }

}
